+++ src/duckylib/utils/vec2.h
@@ -25,5 +25,6 @@ float vec2_magnitude(Vec2 vec);
 Vec2 vec2_normalize(Vec2 vec);
 bool vec2_is_equal(Vec2 a, Vec2 b);
 Vec2 vec2_cross(Vec2 a, Vec2 b);
+float vec2_distance(Vec2 a, Vec2 b);
 
 #endif